Simplified by removing main.cc knowledge from CommandLine
... and removing duplicated storage of options. Parse the originals when
needed.

Also fixed argv[argc] handling: std::vector::reserve() may not zero
"future" items, and it is a bad idea to desynchronize size() and the
number of vector elements we rely on anyway.

Also polished Kid::name*() methods naming. Their internals still need
more work. For example, it does not make sense to add parenthesis and
then remove them. It should be easier to generate the needed name format
when it is needed, possibly using ToSBuf().

Also polished --kid description (without prejudice). May need more work.
TODO: We should update the man page as well.

Also fixed CommandLine.h #include guards that lacked a SQUID_ prefix.

Also fixed Makefile.am order.

Also fixed the last getopt_long() parameter. The original code probably
mistook it for an iteration index. The iteration index is the optind
global.

Also polished the --foreground option ID. The ID of 1 is actually used
for a special getopt_long(3) case. We should avoid it even though we do
not use that special case (yet?).

Except the last two, all the changes refer to the branch code.
